backend/requirements.txt
django 
djangorestframework
markdown
django-filter
python-decouple
django-cors-headers